# faster paths.nim