--- a/src/pixie/blends.nim
+++ b/src/pixie/blends.nim
@@ -160,7 +160,7 @@ proc SetSat(C: Color, s: float32): Color {.inline.} =
   if satC > 0:
     result = (C - min([C.r, C.g, C.b])) * s / satC
 
-proc blendNormal(backdrop, source: ColorRGBA): ColorRGBA =
+proc blendNormal*(backdrop, source: ColorRGBA): ColorRGBA =
   if backdrop.a == 0:
     return source
   if source.a == 255:
@@ -516,6 +516,31 @@ when defined(amd64) and not defined(pixieNoSimd):
     result = mm_or_si128(mm_or_si128(result, i), mm_or_si128(j, k))
     result = mm_and_si128(result, first32)
 
+  proc unpackAlphaValues*(v: M128i): M128i {.inline.} =
+    ## Unpack the first 32 bits into 4 rgba(0, 0, 0, value)
+    let
+      first32 = cast[M128i]([uint32.high, 0, 0, 0]) # First 32 bits
+      alphaMask = mm_set1_epi32(cast[int32](0xff000000)) # Only `a`
+
+    result = mm_shuffle_epi32(v, MM_SHUFFLE(0, 0, 0, 0))
+
+    var
+      i = mm_and_si128(result, first32)
+      j = mm_and_si128(result, mm_slli_si128(first32, 4))
+      k = mm_and_si128(result, mm_slli_si128(first32, 8))
+      l = mm_and_si128(result, mm_slli_si128(first32, 12))
+
+    # Shift the values to `a`
+    i = mm_slli_si128(i, 3)
+    j = mm_slli_si128(j, 2)
+    k = mm_slli_si128(k, 1)
+    # l = mm_slli_si128(l, 0)
+
+    result = mm_and_si128(
+      mm_or_si128(mm_or_si128(i, j), mm_or_si128(k, l)),
+      alphaMask
+    )
+
   proc blendNormalSimd*(backdrop, source: M128i): M128i =
     let
       alphaMask = mm_set1_epi32(cast[int32](0xff000000))
@@ -615,9 +640,7 @@ when defined(amd64) and not defined(pixieNoSimd):
       blendedEven = mm_add_epi16(sourceEven, backdropEven)
       blendedOdd = mm_add_epi16(sourceOdd, backdropOdd)
 
-    blendedOdd = mm_slli_epi16(blendedOdd, 8)
-
-    mm_or_si128(blendedEven, blendedOdd)
+    mm_or_si128(blendedEven, mm_slli_epi16(blendedOdd, 8))
 
   proc maskMaskSimd*(backdrop, source: M128i): M128i =
     let
@@ -638,9 +661,7 @@ when defined(amd64) and not defined(pixieNoSimd):
     backdropEven = mm_srli_epi16(mm_mulhi_epu16(backdropEven, div255), 7)
     backdropOdd = mm_srli_epi16(mm_mulhi_epu16(backdropOdd, div255), 7)
 
-    backdropOdd = mm_slli_epi16(backdropOdd, 8)
-
-    mm_or_si128(backdropEven, backdropOdd)
+    mm_or_si128(backdropEven, mm_slli_epi16(backdropOdd, 8))
 
   proc maskerSimd*(blendMode: BlendMode): MaskerSimd =
     case blendMode: