tiger timing `nim c -d:release --debugger:native -f -r .\tests\benchmark_svg.nim`
before:
`svg decode ........................ 36.423 ms     37.024 ms    ±0.465   x135`
after:
`svg decode ........................ 23.111 ms     24.079 ms    ±0.677   x207`